Rename `convert.base` -> `convert.core`
